-setStrictChooseMode and assertAllCommandsUsed in every test -store filter info in GameState -put instead of putIfAbsent
