+++ Mage/src/main/java/mage/game/Game.java
@@ -17,7 +17,6 @@ import mage.cards.decks.Deck;
 import mage.choices.Choice;
 import mage.constants.*;
 import mage.counters.Counters;
-import mage.filter.common.FilterCreaturePermanent;
 import mage.game.combat.Combat;
 import mage.game.command.Commander;
 import mage.game.command.Emblem;
@@ -486,9 +485,5 @@ public interface Game extends MageItem, Serializable {
     default Set<UUID> getCommandersIds(Player player) {
         return getCommandersIds(player, CommanderCardType.ANY);
     }
-
-    void addPowerInsteadOfToughnessForDamageLethalityFilter(UUID source, FilterCreaturePermanent filter);
-
-    List<FilterCreaturePermanent> getActivePowerInsteadOfToughnessForDamageLethalityFilters();
 
 }
